Give up, use apscheduelr

+++ aaron/worker/scheduler.py
@@ -3,9 +3,29 @@ Scheduled/recurring tasks.
 """
 from __future__ import absolute_import, print_function, unicode_literals
 import json
+import logging
 import os
 
+from apscheduler.schedulers.blocking import BlockingScheduler
+from apscheduler.jobstores.memory import MemoryJobStore
+from apscheduler.executors.pool import ThreadPoolExecutor
 import boto.sqs
+import pytz
+
+logger = logging.basicConfig()
+jobstores = {
+    "default": MemoryJobStore()
+}
+executors = {
+    "default": ThreadPoolExecutor(20)
+}
+job_defaults = {
+    "coalesce": True,
+    "max_instances": 1,
+    "misfire_grace_time": None,  # always execute missed jobs
+}
+scheduler = BlockingScheduler(jobstores=jobstores, executors=executors,
+                              timezone=pytz.utc, logger=logger)
 
 conn = boto.sqs.connect_to_region(
     "us-east-1",
@@ -30,18 +50,26 @@ def run_job(obj):
 
 
 def process_jobs():
-    jobs = input_queue.get_messages(num_messages=10)
+    while True:
+        jobs = input_queue.get_messages(num_messages=10)
 
-    for job in jobs:
-        job_body = job.get_body()
-        print("Received job", job_body)
-        result = run_job(job_body)
+        if not jobs:
+            break
 
-        new_message = boto.sqs.message.Message()
-        new_message.set_body(result)
-        output_queue.write(new_message)
+        for job in jobs:
+            job_body = job.get_body()
+            print("Received job", job_body)
+            result = run_job(job_body)
 
-        print("Sent Job", json.dumps(result))
+            new_message = boto.sqs.message.Message()
+            new_message.set_body(result)
+            output_queue.write(new_message)
 
-        input_queue.delete_message(job)
-    process_jobs()
+            print("Sent Job", json.dumps(result))
+
+            input_queue.delete_message(job)
+
+
+scheduler.add_job(process_jobs, "interval", seconds=1)
+
+scheduler.start()
